+++ src/game/config.ts
@@ -1,8 +1,10 @@
 import { CANVAS, Scale, Types } from 'phaser';
 import { AssetPreviewScene } from '../scenes/AssetPreviewScene';
 import { BootScene } from '../scenes/BootScene';
+import { CafeScene } from '../scenes/CafeScene';
 import { MapScene } from '../scenes/MapScene';
 import { OfficeScene } from '../scenes/OfficeScene';
+import { PoliceKioskScene } from '../scenes/PoliceKioskScene';
 import { PreloadScene } from '../scenes/PreloadScene';
 import { StreetScene } from '../scenes/StreetScene';
 import { TitleScene } from '../scenes/TitleScene';
@@ -24,6 +26,8 @@ export const createGameConfig = (parent: string): Types.Core.GameConfig => ({
         TitleScene,
         OfficeScene,
         StreetScene,
+        CafeScene,
+        PoliceKioskScene,
         MapScene,
         AssetPreviewScene
     ]